+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Create a static context for new traces
+    - Fix memory leak of the staring trace by pregenerate the context before starting a trace (see: [MemoryLeakContextProblem] in the code)
 
 ## 6.5.0 - 2022-06-02
 - Add custom identifier as a tag to custom scoped trace